CompletedTest

+++ CaaS.DTO/AddressDTO.cs
@@ -7,7 +7,7 @@ public class AddressDTO{
     public string Id { get; set; }
     public string Street { get; set; }
     public string Floor { get; set; }
-    public double PostalCode { get; set; }
+    public int PostalCode { get; set; }
     public string City { get; set; }
     public string Province { get; set; }
     public string Country { get; set; }
